\name{sparc-package}
\alias{sparc-package}
\docType{package}
\title{
Semiparametric Generalized Linear Models
}
\description{
We provide an efficient solver for estimating semiparametric generalized linear models.}
\details{
\tabular{ll}{
Package: \tab sparc\cr
Type: \tab Package\cr
Version: \tab 0.9.0\cr
Date: \tab 2013-08-31\cr
License: \tab GPL-2 \cr
}
}
\references{
Y. Ning, Y. Chen, and H. Liu. "High Dimensional Semiparametric Generalized Linear Models", \emph{Technical Report}, 2013.
}
\author{
Tuo Zhao and Han Liu\cr
Maintainers: Tuo Zhao<tourzhao@gmail.com>
}
\seealso{
\code{\link{sparc}}
}